Update for 1.2 and 1.2.1 releases
As Sourceforge is currently very very slow to make uploaded files
available for downloading, rewrite the download instructions so that
GitHub releases are the primary download locations.

Change download buttons to green to match those on github.com.

Release notes are now on the GitHub release pages.
We'll back-populate 1.0 and 1.1 releases there.

Now that there are three separate downloads rather than a download directory
to look in, change the big "Download" button on the homepage to point readers
to the instructions instead.