Suggesting biostars as support forum
This PR is meant as the start of a discussion, in which we suggest to redirect users to biostars (https://www.biostars.org/) for questions and PEBCAK bugs, with we = "the biostars moderators and daily contributors". We can take care of most, if not all, questions, and make sure that genuine bugs get reported properly on GitHub. Related twitter thread: https://twitter.com/jomarnz/status/1070614980435853313

The link I tentatively added in the PR redirects to a new post with the `samtools` tag, although it would solve lots of problems if users first use google.

Tagging a few fellow mods: @dpryan79, @lindenb, @RamRS, @genomax, @kevinblighe (list not meant to be complete)