Improve manpage infrastructure
Ensure the versioned directory for the man pages is created
at the same time as the non-versioned copies.  This makes it easier
to add them to the list of old pages on the next release.

Not complete, as the results need some manual editing - notably
the current version ends up on the list of old versions in
`doc.md`.