Fix manpage style sheet: Make <dl> self-clearing
Not all .SH/.SS headings are preceded by .PP/etc, so we need to ensure
the resulting <h1>/<h2> clears the resulting floats too. Even better,
ensure <dl> clears its own floated <dt> and <dd> children so that the
following text is rendered below the definition list regardless of
what it is. Hat tip @daviesrob.

This makes a difference with Safari, where previously the heading
appeared off to the right of the page alongside the previous .TP list.