Update CRAM benchmark documentation.
This brings the CRAM 3.1 benchmarks up to date with some newer file formats.

It also adds documentation on the 3.0 to 3.1 change in the new release
of htslib.

The old benchmarks included some other things, such as benchmarking
the minhash sort and name sorted data, so they are retained and
renamed (for now).